requirements.txt
streamlit
langchain
langchain-openai
langchain-anthropic
langchain-community
pypdf
tiktoken
python-dotenv
matplotlib
pandas
